+++ nzshm_model/logic_tree/__init__.py
@@ -12,6 +12,7 @@ Classes:
 
 """
 
+from .branch import CompositeBranch
 from .correlation import Correlation, LogicTreeCorrelations
 from .gmcm_logic_tree import GMCMBranch, GMCMBranchSet, GMCMLogicTree
 from .source_logic_tree import InversionSource, SourceBranch, SourceBranchSet, SourceLogicTree, SourceLogicTreeV1