+++ nzshm_model/logic_tree/logic_tree_base.py
@@ -4,7 +4,6 @@ This module contains base classes (some of which are abstract) common to both **
 """
 import copy
 import json
-import math
 from abc import ABC, abstractmethod
 from collections.abc import Sequence
 from dataclasses import asdict, dataclass, field, fields
@@ -12,10 +11,11 @@ from functools import reduce
 from itertools import product
 from operator import mul
 from pathlib import Path
-from typing import Any, Dict, Generator, Iterator, List, Type, Union
+from typing import Any, Dict, Generator, Iterator, List, Type, TypeVar, Union
 
 import dacite
 
+import nzshm_model.logic_tree.helpers as helpers
 from nzshm_model.psha_adapter import PshaAdapterInterface
 
 from .branch import Branch, CompositeBranch
@@ -27,6 +27,9 @@ from .correlation import LogicTreeCorrelations
 #    the initializer of FilteredBranch?
 # - FilteredBranch doesn't need to be a data class as it should not be serialized and doesn't contain many arguments
 # - should we use FilteredBranch for correlation so the branches can be traced back to the BranchSet?
+
+# https://github.com/python/mypy/issues/8495
+LogicTreeType = TypeVar("LogicTreeType", bound="LogicTree")
 
 
 @dataclass
@@ -46,19 +49,7 @@ class BranchSet:
     branches: Sequence[Any] = field(default_factory=list)
 
     def __post_init__(self):
-        if not self._validate_weights():
-            raise ValueError("weights of BranchSet must sum to 1.0")
-
-    def _validate_weights(self) -> bool:
-        """
-        verify that weighs sum to 1.0
-        """
-        weight = 0.0
-        if not self.branches:  # empty BranchSet
-            return True
-        for b in self.branches:
-            weight += b.weight
-        return math.isclose(weight, 1.0)
+        helpers._validate_branchset_weights(self)
 
 
 @dataclass
@@ -82,20 +73,12 @@ class LogicTree(ABC):
     correlations: LogicTreeCorrelations = field(default_factory=LogicTreeCorrelations)
 
     def __post_init__(self) -> None:
-        self._validate_correlations()
+        helpers._validate_correlation_weights(self)
 
     def __setattr__(self, __name: str, __value: Any) -> None:
         super().__setattr__(__name, __value)
         if __name == "correlations":
-            self._validate_correlations()
-
-    def _validate_correlations(self) -> None:
-        # check that the weights total 1.0
-        weight_total = 0.0
-        for branch in self.combined_branches:
-            weight_total += branch.weight
-        if not math.isclose(weight_total, 1.0):
-            raise ValueError("the weights of the logic tree do not sum to 1.0 when correlations are applied")
+            helpers._validate_correlation_weights(self)
 
     def _combined_branches(self) -> Generator[CompositeBranch, None, None]:
         """
@@ -130,7 +113,10 @@ class LogicTree(ABC):
     @classmethod
     def from_json(cls, json_path: Union[Path, str]) -> 'LogicTree':
         """
-        Create LogicTree object from json file or string
+        Create LogicTree object from json file
+
+        See docs/api/logic_tree/source_logic_tree_config_format.md and
+        api/logic_tree/gmcm_logic_tree_config_format.md
 
         Parameters:
             json_path: path to json file
@@ -143,9 +129,36 @@ class LogicTree(ABC):
         return cls.from_dict(data)
 
     @classmethod
-    def from_dict(cls, data: Dict) -> 'LogicTree':
+    def from_dict(cls: Type[LogicTreeType], data: Dict) -> LogicTreeType:
         """
-        Create LogicTree object from dict
+        Create LogicTree object from dict.
+
+        See docs/api/logic_tree/source_logic_tree_config_format.md and
+        api/logic_tree/gmcm_logic_tree_config_format.md
+
+        Parameters:
+            data: dict representation of LogicTree object
+
+        Returns:
+            logic_tree
+        """
+        if not data.get('correlations'):
+            logic_tree = cls._from_dict(data)
+            # do not need to validate names as that is only necessary if there are correlations
+            return logic_tree
+
+        correlations = data.pop('correlations')
+        helpers._validate_correlations_format(correlations)
+        logic_tree = cls._from_dict(data)
+        helpers._validate_names(logic_tree)
+        logic_tree = helpers._add_corellations(logic_tree, correlations)
+
+        return logic_tree
+
+    @classmethod
+    def _from_dict(cls: Type[LogicTreeType], data: Dict) -> LogicTreeType:
+        """
+        Create LogicTree object from dict. Input dict must be a direct asdict() serialisation of the LogicTree object
 
         Parameters:
             data: dict representation of LogicTree object
@@ -155,8 +168,40 @@ class LogicTree(ABC):
         """
         return dacite.from_dict(data_class=cls, data=data, config=dacite.Config(strict=True))
 
-    def to_dict(self) -> Dict[str, Any]:
+    def _to_dict(self) -> Dict[str, Any]:
+        """Create dict representation of logic tree. This creates an exact dict of the class and is not used for serialisation.
+
+        Returns:
+            dict:
+        """
         return asdict(self)
+
+    def to_dict(self) -> Dict[str, Any]:
+        """Create dictionary representation of logic tree used for serialisation.
+
+        Returns:
+            logic_tree_dict: dictionary representaion of logic tree
+        """
+
+        # do not need to validate names as that is only necessary if there are correlations
+        data = self._to_dict()
+        if not self.correlations:
+            del data["correlations"]
+            return data
+
+        helpers._validate_names(self)
+        data["correlations"] = helpers._serialise_correlations(self)
+        return data
+
+    def to_json(self, file_path: Union[Path, str]) -> None:
+        """Serialze logic tree as json file.
+
+        Parameters:
+            file_path: path to json file to be written
+        """
+        file_path = Path(file_path)
+        with file_path.open('w') as jsonfile:
+            json.dump(self.to_dict(), jsonfile)
 
     # would like this to actully do the work, but not sure how to pass the logic trees wihtout knowning the type.
     # Could check for type in PshaAdaptorInterface, but then we have a circular import.