+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.9.2] 2024-02-26
+
+### Added
+ - create LogicTree objects from hand-built json
+
+### Changed
+ - alter form of serialized LogicTree object
+
 ## [0.9.1] 2024-02-26
 
 ### Changed